+++ src/test/java/org/opensearch/knn/integ/KNNScriptScoringIT.java
@@ -71,16 +71,32 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         testKNNScriptScore(SpaceType.L2);
     }
 
+    public void testKNNL2ByteScriptScore() throws Exception {
+        testKNNByteScriptScore(SpaceType.L2);
+    }
+
     public void testKNNL1ScriptScore() throws Exception {
         testKNNScriptScore(SpaceType.L1);
+    }
+
+    public void testKNNL1ByteScriptScore() throws Exception {
+        testKNNByteScriptScore(SpaceType.L1);
     }
 
     public void testKNNLInfScriptScore() throws Exception {
         testKNNScriptScore(SpaceType.LINF);
     }
 
+    public void testKNNLInfByteScriptScore() throws Exception {
+        testKNNByteScriptScore(SpaceType.LINF);
+    }
+
     public void testKNNCosineScriptScore() throws Exception {
         testKNNScriptScore(SpaceType.COSINESIMIL);
+    }
+
+    public void testKNNByteCosineScriptScore() throws Exception {
+        testKNNByteScriptScore(SpaceType.COSINESIMIL);
     }
 
     @SneakyThrows
@@ -88,16 +104,21 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         testKNNScriptScoreOnBinaryIndex(SpaceType.HAMMING);
     }
 
+    @SuppressWarnings("unchecked")
     @SneakyThrows
     public void testKNNHammingScriptScore_whenNonBinary_thenException() {
         final int dims = randomIntBetween(2, 10) * 8;
         final float[] queryVector = randomVector(dims, VectorDataType.BYTE);
-        final BiFunction<float[], float[], Float> scoreFunction = getScoreFunction(SpaceType.HAMMING, queryVector);
+        final BiFunction<byte[], byte[], Float> scoreFunction = (BiFunction<byte[], byte[], Float>) getScoreFunction(
+            SpaceType.HAMMING,
+            queryVector,
+            VectorDataType.BINARY
+        );
         List<VectorDataType> nonBinary = List.of(VectorDataType.FLOAT, VectorDataType.BYTE);
         for (VectorDataType vectorDataType : nonBinary) {
             Exception e = expectThrows(
                 Exception.class,
-                () -> createIndexAndAssertScriptScore(
+                () -> createIndexAndAssertByteScriptScore(
                     createKnnIndexMapping(FIELD_NAME, dims, vectorDataType),
                     SpaceType.HAMMING,
                     scoreFunction,
@@ -112,15 +133,20 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         }
     }
 
+    @SuppressWarnings("unchecked")
     public void testKNNNonHammingScriptScore_whenBinary_thenException() {
         final int dims = randomIntBetween(2, 10) * 8;
         final float[] queryVector = randomVector(dims, VectorDataType.BINARY);
-        final BiFunction<float[], float[], Float> scoreFunction = getScoreFunction(SpaceType.HAMMING, queryVector);
+        final BiFunction<byte[], byte[], Float> scoreFunction = (BiFunction<byte[], byte[], Float>) getScoreFunction(
+            SpaceType.HAMMING,
+            queryVector,
+            VectorDataType.BINARY
+        );
         Set<SpaceType> spaceTypeToExclude = Set.of(SpaceType.UNDEFINED, SpaceType.HAMMING);
         Arrays.stream(SpaceType.values()).filter(s -> spaceTypeToExclude.contains(s) == false).forEach(s -> {
             Exception e = expectThrows(
                 Exception.class,
-                () -> createIndexAndAssertScriptScore(
+                () -> createIndexAndAssertByteScriptScore(
                     createKnnIndexMapping(FIELD_NAME, dims, VectorDataType.BINARY),
                     s,
                     scoreFunction,
@@ -627,6 +653,7 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         assertEquals(1, secondQueryCacheMap.get("hit_count"));
     }
 
+    @SuppressWarnings("unchecked")
     public void testKNNScriptScoreOnModelBasedIndex() throws Exception {
         int dimensions = randomIntBetween(2, 10);
         String trainMapping = createKnnIndexMapping(TRAIN_FIELD_PARAMETER, dimensions);
@@ -663,7 +690,11 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
                 continue;
             }
             final float[] queryVector = randomVector(dimensions);
-            final BiFunction<float[], float[], Float> scoreFunction = getScoreFunction(spaceType, queryVector);
+            final BiFunction<float[], float[], Float> scoreFunction = (BiFunction<float[], float[], Float>) getScoreFunction(
+                spaceType,
+                queryVector,
+                VectorDataType.FLOAT
+            );
             createIndexAndAssertScriptScore(testMapping, spaceType, scoreFunction, dimensions, queryVector, true);
         }
     }
@@ -686,6 +717,30 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
                 KNNEngine.LUCENE.getName(),
                 SpaceType.DEFAULT.getValue(),
                 false
+            )
+        );
+    }
+
+    private List<String> createByteMappers(int dimensions) throws Exception {
+        return List.of(
+            createKnnIndexMapping(FIELD_NAME, dimensions, VectorDataType.BYTE),
+            createKnnIndexMapping(
+                FIELD_NAME,
+                dimensions,
+                KNNConstants.METHOD_HNSW,
+                KNNEngine.LUCENE.getName(),
+                SpaceType.DEFAULT.getValue(),
+                true,
+                VectorDataType.BYTE
+            ),
+            createKnnIndexMapping(
+                FIELD_NAME,
+                dimensions,
+                KNNConstants.METHOD_HNSW,
+                KNNEngine.LUCENE.getName(),
+                SpaceType.DEFAULT.getValue(),
+                false,
+                VectorDataType.BYTE
             )
         );
     }
@@ -747,7 +802,7 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         return dataset;
     }
 
-    private BiFunction<float[], float[], Float> getScoreFunction(SpaceType spaceType, float[] queryVector) {
+    private BiFunction<?, ?, Float> getScoreFunction(SpaceType spaceType, float[] queryVector, VectorDataType vectorDataType) {
         List<Float> target = new ArrayList<>(queryVector.length);
         for (float f : queryVector) {
             target.add(f);
@@ -758,8 +813,8 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
             new KNNVectorFieldType(
                 FIELD_NAME,
                 Collections.emptyMap(),
-                SpaceType.HAMMING == spaceType ? VectorDataType.BINARY : VectorDataType.FLOAT,
-                getMappingConfigForFlatMapping(SpaceType.HAMMING == spaceType ? queryVector.length * 8 : queryVector.length)
+                vectorDataType,
+                getMappingConfigForFlatMapping(vectorDataType == VectorDataType.BINARY ? queryVector.length * 8 : queryVector.length)
             )
         );
         switch (spaceType) {
@@ -769,35 +824,63 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
             case COSINESIMIL:
             case INNER_PRODUCT:
             case HAMMING:
-                return ((KNNScoringSpace.KNNFieldSpace) knnScoringSpace).getScoringMethod();
+                if (vectorDataType == VectorDataType.FLOAT) {
+                    return ((KNNScoringSpace.KNNFieldSpace) knnScoringSpace).getScoringMethod(queryVector);
+                }
+                return ((KNNScoringSpace.KNNFieldSpace) knnScoringSpace).getScoringMethod(toByte(queryVector));
             default:
                 throw new IllegalArgumentException();
         }
     }
 
+    @SuppressWarnings("unchecked")
     private void testKNNScriptScore(SpaceType spaceType) throws Exception {
         final int dims = randomIntBetween(2, 10);
         final float[] queryVector = randomVector(dims);
-        final BiFunction<float[], float[], Float> scoreFunction = getScoreFunction(spaceType, queryVector);
+        final BiFunction<float[], float[], Float> scoreFunction = (BiFunction<float[], float[], Float>) getScoreFunction(
+            spaceType,
+            queryVector,
+            VectorDataType.FLOAT
+        );
         for (String mapper : createMappers(dims)) {
             createIndexAndAssertScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, true);
             createIndexAndAssertScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, false);
         }
     }
 
+    @SuppressWarnings("unchecked")
+    private void testKNNByteScriptScore(SpaceType spaceType) throws Exception {
+        final int dims = randomIntBetween(2, 10);
+        final float[] queryVector = randomVector(dims, VectorDataType.BYTE);
+        final BiFunction<byte[], byte[], Float> scoreFunction = (BiFunction<byte[], byte[], Float>) getScoreFunction(
+            spaceType,
+            queryVector,
+            VectorDataType.BYTE
+        );
+        for (String mapper : createByteMappers(dims)) {
+            createIndexAndAssertByteScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, true, true, VectorDataType.BYTE);
+            createIndexAndAssertByteScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, false, true, VectorDataType.BYTE);
+        }
+    }
+
+    @SuppressWarnings("unchecked")
     private void testKNNScriptScoreOnBinaryIndex(SpaceType spaceType) throws Exception {
         final int dims = randomIntBetween(2, 10) * 8;
         final float[] queryVector = randomVector(dims, VectorDataType.BINARY);
-        final BiFunction<float[], float[], Float> scoreFunction = getScoreFunction(spaceType, queryVector);
+        final BiFunction<byte[], byte[], Float> scoreFunction = (BiFunction<byte[], byte[], Float>) getScoreFunction(
+            spaceType,
+            queryVector,
+            VectorDataType.BINARY
+        );
 
         // Test when knn is enabled and engine is Faiss
         for (String mapper : createBinaryIndexMappers(dims)) {
-            createIndexAndAssertScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, true, true, VectorDataType.BINARY);
-            createIndexAndAssertScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, false, true, VectorDataType.BINARY);
+            createIndexAndAssertByteScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, true, true, VectorDataType.BINARY);
+            createIndexAndAssertByteScriptScore(mapper, spaceType, scoreFunction, dims, queryVector, false, true, VectorDataType.BINARY);
         }
 
         // Test when knn is disabled and engine is default(Nmslib)
-        createIndexAndAssertScriptScore(
+        createIndexAndAssertByteScriptScore(
             createKnnIndexMapping(FIELD_NAME, dims, VectorDataType.BINARY),
             spaceType,
             scoreFunction,
@@ -807,7 +890,7 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
             false,
             VectorDataType.BINARY
         );
-        createIndexAndAssertScriptScore(
+        createIndexAndAssertByteScriptScore(
             createKnnIndexMapping(FIELD_NAME, dims, VectorDataType.BINARY),
             spaceType,
             scoreFunction,
@@ -827,13 +910,22 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         float[] queryVector,
         boolean dense
     ) throws Exception {
-        createIndexAndAssertScriptScore(mapper, spaceType, scoreFunction, dimensions, queryVector, dense, true, VectorDataType.FLOAT);
+        createIndexAndAssertScriptScore(
+            mapper,
+            spaceType,
+            v -> scoreFunction.apply(queryVector, v),
+            dimensions,
+            queryVector,
+            dense,
+            true,
+            VectorDataType.FLOAT
+        );
     }
 
     private void createIndexAndAssertScriptScore(
         String mapper,
         SpaceType spaceType,
-        BiFunction<float[], float[], Float> scoreFunction,
+        Function<float[], Float> scoreFunction,
         int dimensions,
         float[] queryVector,
         boolean dense,
@@ -857,13 +949,7 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         createKnnIndex(INDEX_NAME, settings, mapper);
         try {
             final int numDocsWithField = randomIntBetween(4, 10);
-            Map<String, KNNResult> dataset = createDataset(
-                v -> scoreFunction.apply(queryVector, v),
-                dimensions,
-                numDocsWithField,
-                dense,
-                vectorDataType
-            );
+            Map<String, KNNResult> dataset = createDataset(scoreFunction, dimensions, numDocsWithField, dense, vectorDataType);
             float[] dummyVector = new float[1];
             dataset.forEach((k, v) -> {
                 final float[] vector = (v != null) ? v.getVector() : dummyVector;
@@ -893,6 +979,36 @@ public class KNNScriptScoringIT extends KNNRestTestCase {
         } finally {
             deleteKNNIndex(INDEX_NAME);
         }
+    }
+
+    private void createIndexAndAssertByteScriptScore(
+        String mapper,
+        SpaceType spaceType,
+        BiFunction<byte[], byte[], Float> scoreFunction,
+        int dimensions,
+        float[] queryVector,
+        boolean dense,
+        boolean enableKnn,
+        VectorDataType vectorDataType
+    ) throws Exception {
+        createIndexAndAssertScriptScore(
+            mapper,
+            spaceType,
+            v -> scoreFunction.apply(toByte(queryVector), toByte(v)),
+            dimensions,
+            queryVector,
+            dense,
+            enableKnn,
+            vectorDataType
+        );
+    }
+
+    private byte[] toByte(final float[] vector) {
+        byte[] bytes = new byte[vector.length];
+        for (int i = 0; i < vector.length; i++) {
+            bytes[i] = (byte) vector[i];
+        }
+        return bytes;
     }
 
     private float[] dummyFloatArrayBasedOnDimension(int dimesion) {